feat(mcp): ship structured tool results once by default, text fallback opt-in
The spec's duplicate text copy is a backwards-compatibility SHOULD that
doubles every typed result on the wire; size-capped clients count both
copies and reject payloads at double their real size. textFallback now
defaults to false — servers whose clients ignore outputSchema opt in
with { textFallback: true }. Void results, validation failures, and
runtime errors keep their text content as before.
